% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_style.R
\name{theme_hm_pres}
\alias{theme_hm_pres}
\title{Add hm theme to ggplot chart for presentations}
\usage{
theme_hm_pres()
}
\description{
This function allows you to add the hm theme to your ggplot graphics.
}
\examples{
ldefault_plot <- ggplot(economics, aes(date, unemploy)) +
   geom_line() +
     labs(
         title = "Unemployed persons in the United States",
         subtitle = "Monthly aggregation from 1967 - 2015",
         caption = "Data source: {ggplot2} R package",
         x = NULL,
         y = NULL
     ) +
     # Unemployment level is number of unemployed in thousands.
     # This declutters the figure by presenting the number of
     # unemployed in millions
     scale_y_continuous(
         labels = scales::label_number(
             scale = 0.001,
             suffix = " M"
         ),
         limits = c(0, max(economics$unemploy))
     ) +
theme_hm_pres()
}
\keyword{hm_style}
