% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_style.R
\name{theme_hm_pres}
\alias{theme_hm_pres}
\title{Add hm theme to ggplot chart for presentations}
\usage{
theme_hm_pres()
}
\description{
This function allows you to add the hm theme to your ggplot graphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
hm_style()
}
\keyword{hm_style}
