% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_style.R
\name{theme_hm_pub}
\alias{theme_hm_pub}
\title{Add hm theme to ggplot chart for publications}
\usage{
theme_hm_pub()
}
\value{
A ggplot2 theme object.
}
\description{
As [theme_hm_pres()] but with text sized for a journal figure.
}
\examples{
library(ggplot2)
ggplot(economics, aes(date, unemploy)) +
  geom_line() +
  labs(
    title = "Unemployed persons in the United States",
    subtitle = "Monthly aggregation from 1967 - 2015",
    caption = "Data source: {ggplot2} R package",
    x = NULL,
    y = NULL
  ) +
  theme_hm_pub()
}
\seealso{
[theme_hm_pres()] for the larger presentation sizing.
}
\keyword{hm_style}
